♻️ - Rename example middleware.ts to proxy.ts (Next 16)
Next 16 deprecates the `middleware` file convention in favour of `proxy`.
Rename src/middleware.ts -> src/proxy.ts and the exported `middleware`
function -> `proxy` across all three Next.js examples (nextjs, nextjs-vercel,
nextjs-rsc), and update the docs/comments accordingly.
