+++ examples/nextjs-rsc/src/app/getVariant.ts
@@ -3,7 +3,7 @@ import { cookies } from 'next/headers'
 import { AB_TEST_SLUG, Variant } from './improveConfig'
 
 /**
- * Reads the variant the middleware assigned for this visitor.
+ * Reads the variant the proxy assigned for this visitor.
  *
  * `cookies()` is a dynamic API, so any component that calls this must be
  * rendered inside a <Suspense> boundary when Cache Components is enabled.
