+++ examples/nextjs-vercel/src/proxy.ts
@@ -1,6 +1,8 @@
+import { type NextRequest } from 'next/server'
+import { get as getFromEdgeConfig } from '@vercel/edge-config'
+import { type ImproveConfiguration } from '@obelism/improve-sdk/types'
 import { ImproveServerSDK } from '@obelism/improve-sdk/server'
 import { generateImproveNextMiddleware } from '@obelism/improve-sdk-next'
-import { type NextRequest } from 'next/server'
 
 import { IMPROVE_CONFIG } from './app/improveConfig'
 
@@ -33,7 +35,19 @@ const improveMiddlewareHandler = generateImproveNextMiddleware({
 	],
 })
 
-export const middleware = async (request: NextRequest) => {
-	await improveSdk.fetchConfig()
+const getImproveConfig = async () => {
+	if (!!improveSdk.config) return
+
+	// Only the active state is saved in edge config
+	if (IMPROVE_CONFIG.state !== 'active') {
+		return await improveSdk.fetchConfig()
+	}
+
+	const config = await getFromEdgeConfig(IMPROVE_CONFIG.environment)
+	improveSdk.loadConfig(config as ImproveConfiguration)
+}
+
+export const proxy = async (request: NextRequest) => {
+	await getImproveConfig()
 	return improveMiddlewareHandler(request)
 }
